support cross-validation; add linear regression (not yet tested); add linear normalizer (Laura); fix a minor bug with LambdaMART; add parse representation of feature vector (no benefit observed); rearrangement of some classes

+++ src/ciir/umass/edu/learning/RankList.java
@@ -62,7 +62,7 @@ public class RankList {
 		rl.remove(k);
 	}
 	
-	public RankList getRanking(int fid)
+	public RankList getRanking(short fid)
 	{
 		double[] score = new double[rl.size()];
 		for(int i=0;i<rl.size();i++)
